refactor(theme): extract asset pipeline
